[Lib] Adapt sources for dependencies on data structures
