+++ libs/librrgraph/src/base/metadata_storage.h
